refactor: reorganize project structure and update package references for core model
