+++ .idea/gradle.xml
@@ -23,10 +23,6 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/auth" />
-            <option value="$PROJECT_DIR$/auth/data" />
-            <option value="$PROJECT_DIR$/auth/domain" />
-            <option value="$PROJECT_DIR$/auth/presentation" />
             <option value="$PROJECT_DIR$/build-logic" />
             <option value="$PROJECT_DIR$/build-logic/convention" />
             <option value="$PROJECT_DIR$/content" />
