--- auth/presentation/build.gradle.kts
@@ -1,16 +1,0 @@
-plugins {
-  alias(libs.plugins.application.android.feature.ui)
-}
-
-android {
-  namespace = "com.example.auth.presentation"
-}
-
-dependencies {
-  testImplementation(libs.junit)
-  androidTestImplementation(libs.androidx.junit)
-  androidTestImplementation(libs.androidx.espresso.core)
-
-  implementation(projects.core.domain)
-  implementation(projects.auth.domain)
-}
