+++ content/presentation/src/main/java/com/example/content/presentation/details/DetailsState.kt
@@ -1,6 +1,6 @@
 package com.example.content.presentation.details
 
-import com.example.core.domain.content.PokemonDetails
+import com.example.core.model.PokemonDetails
 
 data class DetailsState(
   val state: PokemonDetails? = null,
