+++ content/presentation/src/main/java/com/example/content/presentation/home/HomeAction.kt
@@ -1,6 +1,6 @@
 package com.example.content.presentation.home
 
-import com.example.core.domain.content.Pokemon
+import com.example.core.model.Pokemon
 
 sealed interface HomeAction {
   data object FetchPokemons : HomeAction
