+++ content/presentation/src/main/java/com/example/content/presentation/home/mapper/pokemonUiMapper.kt
@@ -1,7 +1,7 @@
 package com.example.content.presentation.home.mapper
 
 import com.example.content.presentation.home.model.PokemonUi
-import com.example.core.domain.content.Pokemon
+import com.example.core.model.Pokemon
 
 fun Pokemon.toPokemonUi(): PokemonUi {
   return PokemonUi(
