+++ core/data/src/main/java/com/example/core/data/content/OfflineFirstPokeRepository.kt
@@ -2,10 +2,10 @@ package com.example.core.data.content
 
 import com.example.core.database.RoomLocalPokemonDataSource
 import com.example.core.domain.content.PokeRepository
-import com.example.core.domain.content.Pokemon
-import com.example.core.domain.content.PokemonDetails
 import com.example.core.domain.content.RemotePokemonDataSource
 import com.example.core.domain.util.Result
+import com.example.core.model.Pokemon
+import com.example.core.model.PokemonDetails
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
