+++ core/domain/src/main/java/com/example/core/domain/content/PokeRepository.kt
@@ -1,5 +1,7 @@
 package com.example.core.domain.content
 
+import com.example.core.model.Pokemon
+import com.example.core.model.PokemonDetails
 import kotlinx.coroutines.flow.Flow
 
 interface PokeRepository {
