+++ core/model/src/main/java/com/example/core/model/PokemonDetails.kt
@@ -1,4 +1,4 @@
-package com.example.core.domain.content
+package com.example.core.model
 
 import kotlin.random.Random
