+++ settings.gradle.kts
@@ -23,17 +23,14 @@ dependencyResolutionManagement {
 rootProject.name = "Application"
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 include(":app")
-include(":auth:data")
-include(":auth:domain")
-include(":auth:presentation")
 include(":core:presentation:designsystem")
 include(":core:presentation:ui")
 include(":core:domain")
 include(":core:data")
 include(":core:database")
+include(":core:navigation")
+include(":core:model")
 include(":content:data")
 include(":content:domain")
 include(":content:presentation")
 include(":content:network")
-include(":core:navigation")
-include(":core:model")
